Starting to implement rendering system
After a lot of testing to make sure that this would work, I decided to
implement the GUI rendering by rendering all components to a separate
render target, and then drawing that render target at the end of the
frame over everything else. Also, it uses the stencil buffer to allow
parent widgets to mask their children.